added lots and lots of documentation to most of the classes in TheRestlessTombs

+++ TheRestlessTombs/healthPotion.cpp
@@ -1,7 +1,9 @@
 #include "healthPotion.h"
 
 HealthPotion::HealthPotion(int healing, Camera* camera, Shader* shader, b2World* world) : Potion::Potion(camera, shader, world) {
+	// Intialize the healing with the healing from the constructor
 	this->healing = healing;
+	// Name the Item "health potion"
 	name = "health potion";
 }
 
@@ -9,6 +11,7 @@ HealthPotion::~HealthPotion() {
 
 }
 
-float HealthPotion::Use() {
+int HealthPotion::Use() {
+	// return the healing
 	return healing;
 }
